EKF move earth spin rate to geo and update usage

+++ EKF/ekf.cpp
@@ -309,7 +309,7 @@ void Ekf::predictState()
 {
 	if (!_earth_rate_initialised) {
 		if (_NED_origin_initialised) {
-			calcEarthRateNED(_earth_rate_NED, _pos_ref.lat_rad);
+			calcEarthRateNED(_earth_rate_NED, (float)_pos_ref.lat_rad);
 			_earth_rate_initialised = true;
 		}
 	}
